(#17) refactor: Auth -> AuthWrapper로 파일이름 변경
