+++ src/pages/auth/SignIn.tsx
@@ -2,7 +2,7 @@ import { LogoWithText } from '@/assets';
 import { Button } from '@/components/common/button';
 import { Input } from '@/components/common/input';
 import { Text } from '@/components/common/text';
-import { AuthWrapper } from '@/layouts/Auth';
+import { AuthWrapper } from '@/layouts/AuthWrapper';
 import { useSignIn } from '@/hooks/useAuth';
 import { useForm } from '@/hooks/useForm';
 import { Link } from 'react-router-dom';
