+++ src/pages/auth/SignUp.tsx
@@ -1,10 +1,11 @@
 import { Button } from '@/components/common/button';
 import { Input } from '@/components/common/input';
-import { AuthWrapper } from '@/layouts/Auth';
+import { AuthWrapper } from '@/layouts/AuthWrapper';
 import { useSignUp } from '@/hooks/useAuth';
 import { useForm } from '@/hooks/useForm';
 import { Text } from '@/components/common/text';
 import { Link } from 'react-router-dom';
+import { RowGap } from '@/layouts/RowGap';
 
 /** 프로필, 이메일 인증 추가 요망 */
 export const SignUpPage = () => {
@@ -20,36 +21,34 @@ export const SignUpPage = () => {
   return (
     <AuthWrapper onSubmit={() => signUpMutate.mutate()}>
       <Text size="heading1">Sign up</Text>
-      <Input
-        value={text.name}
-        margin="30px 0 20px 0"
-        name="name"
-        label="이름"
-        placeholder="이름을 입력해주세요."
-      />
-      <Input
-        value={text.email}
-        margin="15px 0 20px 0"
-        name="email"
-        label="이메일"
-        placeholder="이메일을 입력해주세요."
-      />
-      <Input
-        value={text.password}
-        margin="15px 0 20px 0"
-        name="password"
-        type="password"
-        label="비밀번호"
-        placeholder="비밀번호를 입력해주세요."
-      />
-      <Input
-        value={text.passwordCheck}
-        margin="15px 0 30px 0"
-        name="passwordCheck"
-        type="password"
-        label="비밀번호 확인"
-        placeholder="비밀번호를 한 번 더 입력해주세요."
-      />
+      <RowGap gap="20px">
+        <Input
+          value={text.name}
+          name="name"
+          label="이름"
+          placeholder="이름을 입력해주세요."
+        />
+        <Input
+          value={text.email}
+          name="email"
+          label="이메일"
+          placeholder="이메일을 입력해주세요."
+        />
+        <Input
+          value={text.password}
+          name="password"
+          type="password"
+          label="비밀번호"
+          placeholder="비밀번호를 입력해주세요."
+        />
+        <Input
+          value={text.passwordCheck}
+          name="passwordCheck"
+          type="password"
+          label="비밀번호 확인"
+          placeholder="비밀번호를 한 번 더 입력해주세요."
+        />
+      </RowGap>
       <Button>회원가입</Button>
       <Text margin="20px 0 0 0" align="center" size="body1">
         <Link to="/signin">로그인</Link> |
